% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-plotly.R
\name{abm_bullet_plotly}
\alias{abm_bullet_plotly}
\title{Create alternate bullet graph with reference line (using ggplot conversion to plotly)}
\usage{
abm_bullet_plotly(label, value, reference, roundto = 1, pct = FALSE)
}
\arguments{
\item{label}{a label for the indicator, shown to the left of the gauge}

\item{value}{the value of the indicator, displayed as a horizontal wide line}

\item{reference}{a reference value displayed as a vertical thin line}

\item{roundto}{number of digits after the decimal point (default = 1)}

\item{pct}{boolean, set to TRUE if given value is a share (default = FALSE)}
}
\value{
a ggplot object
}
\description{
Create alternate bullet graph with reference line (using ggplot conversion to plotly)
}
