% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-plotly.R
\name{abm_bullet_plotly2}
\alias{abm_bullet_plotly2}
\title{Create alternate bullet graph with reference line (using plotly directly)}
\usage{
abm_bullet_plotly2(
  caption = "Field normalized citations (Cf) = 1.13",
  value = 1.13,
  reference = 1,
  span = c(0, 2)
)
}
\arguments{
\item{caption}{a label for the indicator, shown to the left of the gauge}

\item{value}{the value of the indicator, displayed as a horizontal wide line}

\item{reference}{a reference value displayed as a vertical thin line}

\item{span}{number of digits after the decimal point (default = 1)}
}
\value{
a plotly object
}
\description{
Create alternate bullet graph with reference line (using plotly directly)
}
