% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_config}
\alias{abm_config}
\title{Provide defaults for ABM analysis}
\usage{
abm_config()
}
\value{
list of default values used in the current abm presentation
}
\description{
Provide defaults for ABM analysis
}
