% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_copub_countries}
\alias{abm_copub_countries}
\title{Create table over co-publication countries for ABM unit}
\usage{
abm_copub_countries(
  con,
  analysis_id = abm_config()$analysis_id,
  unit_level,
  unit_code,
  exclude_swe = TRUE,
  limit = NULL,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{con}{a database connection to BIBMON}

\item{analysis_id}{id for the analysis, default from abm_config()}

\item{unit_level}{organization level}

\item{unit_code}{code for the analyzed unit}

\item{exclude_swe}{wether to exclude Sweden as co-publication country, default TRUE}

\item{limit}{if set, limit the result to the first limit rows, default NULL}

\item{analysis_start}{first publication year of analysis, default from abm_config()}

\item{analysis_stop}{last publication year of analysis, default from abm_config()}
}
\value{
a tibble
}
\description{
Create table over co-publication countries for ABM unit
}
