% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_copub_data}
\alias{abm_copub_data}
\title{Retrieve co-publishing organizations for ABM tables}
\usage{
abm_copub_data(
  con = con_bib(),
  unit_code,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{for filtering on one or more unit code(s), which can be KTH, a one letter school code, an integer department code or a KTH-id (optional)}

\item{analysis_start}{first publication year of analysis, default 2012}

\item{analysis_stop}{last publication year of analysis, default 2018}
}
\value{
tibble with co-publishing organizations associated with each publication for the selected ABM-organization
}
\description{
This function retrieves all co-publishing organizations for the selected ABM-unit, for all
publications that has a UT-number (WebofScience_id). The returned tibble has one row per organization 
and publication.
}
