% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_copub_orgs}
\alias{abm_copub_orgs}
\title{Create table over co-publication countries for ABM unit}
\usage{
abm_copub_orgs(
  con,
  analysisId = abm_config()$analysis_id,
  unit_level,
  unit_code,
  exclude_swe = FALSE,
  limit = 1000,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{con}{a database connection to BIBMON}

\item{analysisId}{id for the analysis, default from abm_config()}

\item{unit_level}{organization level}

\item{unit_code}{code for the analyzed unit}

\item{exclude_swe}{wether to exclude Swedish co-publication orgs, default FALSE}

\item{limit}{if set, limit the result (for level 0 and 1) to the first limit rows, default 1000}

\item{analysis_start}{first publication year of analysis, default from abm_config()}

\item{analysis_stop}{last publication year of analysis, default from abm_config()}
}
\value{
a tibble
}
\description{
Create table over co-publication countries for ABM unit
}
