% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_staffbased.R
\name{abm_copub_orgs_staffbased}
\alias{abm_copub_orgs_staffbased}
\title{Create staffbased table over co-publication organizations for ABM unit}
\usage{
abm_copub_orgs_staffbased(con, publist, exclude_swe = FALSE, limit = 1000)
}
\arguments{
\item{con}{a database connection to BIBMON}

\item{publist}{a publication list including at least WebofScience_ID, Unit_fraction and n_authors}

\item{exclude_swe}{wether to exclude Sweden as co-publication country, default TRUE}

\item{limit}{if set, limit the result to the first limit rows, default NULL}
}
\value{
a tibble
}
\description{
Create staffbased table over co-publication organizations for ABM unit
}
