% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_coverage}
\alias{abm_coverage}
\title{Retrieve WoS and Scopus coverage for peer reviewed DiVA publication types}
\usage{
abm_coverage(
  data,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{data}{dataset with publications as tibble}

\item{analysis_start}{first publication year of analysis, if not given abm_config() is used}

\item{analysis_stop}{last publication year of analysis, if not given abm_config() is used}
}
\value{
tibble with fractionalized and full counted WoS coverage by year and publication type
}
\description{
Retrieve WoS and Scopus coverage for peer reviewed DiVA publication types
}
