% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_data}
\alias{abm_data}
\title{Retrieve data for ABM tables and graphs from master table}
\usage{
abm_data(con = con_bib(), unit_code, pub_year, unit_level, analysis_id)
}
\arguments{
\item{con}{connection to db, default is sqlite connection}

\item{unit_code}{for filtering on one or more unit code(s), which can be KTH, a one letter school code, an integer department code or a KTH-id (optional)}

\item{pub_year}{for filtering on publication years, for example 2012, 2012:2018 or c(2012, 2014, 2016) (optional)}

\item{unit_level}{for filtering on organizational level, 0 = KTH, 1 = school, 2 = deparment, 3 = researcher.}

\item{analysis_id}{for filtering on specific analysis, for example 1 = KTH ABM 2020}
}
\value{
tibble with all ABM data for selected organizational units
}
\description{
Retrieve data for ABM tables and graphs from master table
}
