% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_data_workbook}
\alias{abm_data_workbook}
\title{Create Excel workbook with ABM publication data}
\usage{
abm_data_workbook(data, attribution = NULL)
}
\arguments{
\item{data}{a data frame}

\item{attribution}{(optional) attribution text}
}
\value{
A workbook object
}
\description{
Create Excel workbook with ABM publication data
}
