% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kth_oidc.R
\name{abm_decode_jwt}
\alias{abm_decode_jwt}
\title{Decode JWT string}
\usage{
abm_decode_jwt(x)
}
\arguments{
\item{x}{the jwt in Azure KTH OIDC format}
}
\value{
list with relevant slots populated
}
\description{
Decode a Java Web Token string in the format used in Azure (KTH OIDC)
}
\examples{
\dontrun{
if(interactive()){
  abm_decode_jwt(Sys.getenv("SHINYPROXY_OIDC_ACCESS_TOKEN"))
 }
}
}
