% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{abm_division_stats}
\alias{abm_division_stats}
\title{Publication summary stats for ABM divisions}
\usage{
abm_division_stats(slugs = abm_divisions()$id)
}
\arguments{
\item{slugs}{a vector of unit slug identifiers, Default: abm_divisions()$id}
}
\value{
tibble with summary data
}
\description{
Publication summary stats for ABM divisions
}
\examples{
\dontrun{
if(interactive()){
 stats <- abm_division_stats() 
 # upload results to database
 db_upsert_table("division_stats", stats)
 }
}
}
