% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{abm_divisions}
\alias{abm_divisions}
\title{abm_divisions}
\usage{
abm_divisions(
  include = abm_slugs_institutions(),
  exclude = NULL,
  quiet = FALSE
)
}
\arguments{
\item{include}{character vector of slugs to include, Default: abm_slugs_institutions()}

\item{exclude}{character vector of slugs to exclude, Default: NULL}

\item{quiet}{logical to indicate logging, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Divisions used in ABM
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 abm_divisions()
 }
}
}
\seealso{
\code{\link[progress]{progress_bar}}
 \code{\link[purrr]{safely}},\code{\link[purrr]{map}}
}
