% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_format_columns_divatable}
\alias{abm_format_columns_divatable}
\title{Make ABM diva table have last 2 columns with gray background, penultimate with dark gray left border}
\usage{
abm_format_columns_divatable(t, column_name, has_left_border)
}
\arguments{
\item{t}{the table to be formatted}

\item{column_name}{the name of the column to be formatted among c("P_frac","WoS_coverage")}

\item{has_left_border}{boolean to indicate if the column should have a left border}
}
\description{
Make ABM diva table have last 2 columns with gray background, penultimate with dark gray left border
}
