% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_format_header_divatable}
\alias{abm_format_header_divatable}
\title{Make ABM diva table header have last 2 columns with gray background, penultimate with dark gray left border}
\usage{
abm_format_header_divatable(header)
}
\arguments{
\item{header}{the diva table header to be formatted}
}
\description{
Make ABM diva table header have last 2 columns with gray background, penultimate with dark gray left border
}
