% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_format_rows}
\alias{abm_format_rows}
\title{Make ABM table have last rows bold with gray background, other rows with white background}
\usage{
abm_format_rows(t)
}
\arguments{
\item{t}{the table to be formatted}
}
\description{
Make ABM table have last rows bold with gray background, other rows with white background
}
