% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_cf}
\alias{abm_graph_cf}
\title{Create graph over Cf by year}
\usage{
abm_graph_cf(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_table3()}
}
\value{
a ggplot object
}
\description{
Create graph over Cf by year
}
