% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_copub}
\alias{abm_graph_copub}
\title{Create graph over international and Swedish non-university copublications by year}
\usage{
abm_graph_copub(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_table4()}
}
\value{
a ggplot object
}
\description{
Create graph over international and Swedish non-university copublications by year
}
