% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-plotly.R
\name{abm_graph_copub_plotly}
\alias{abm_graph_copub_plotly}
\title{Create graph over international and Swedish non-university copublications by year, using plotly}
\usage{
abm_graph_copub_plotly(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_table4()}
}
\value{
a plotly object
}
\description{
Create graph over international and Swedish non-university copublications by year, using plotly
}
