% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_diva}
\alias{abm_graph_diva}
\title{Create graph over DiVA publication types by year}
\usage{
abm_graph_diva(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_table1()}
}
\value{
a ggplot object
}
\description{
Create graph over DiVA publication types by year
}
