% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-plotly.R
\name{abm_graph_diva2}
\alias{abm_graph_diva2}
\title{Create graph over DiVA publication types by year (ggplot, plotly and ggiraph variants)}
\usage{
abm_graph_diva2(df, type = c("ggplot", "plotly", "ggiraph"))
}
\arguments{
\item{df}{a data frame at the format produced by abm_table1()}

\item{type}{one of ggplot, plotly or ggiraph, where the first one is default}
}
\value{
a ggplot object
}
\description{
Create graph over DiVA publication types by year (ggplot, plotly and ggiraph variants)
}
