% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{abm_graph_divisions}
\alias{abm_graph_divisions}
\title{ABM divisions as force directed network}
\usage{
abm_graph_divisions(base_url = "dash/")
}
\arguments{
\item{base_url}{pattern to prefigate links from nodes with in JS click action}
}
\value{
a force directed network object from NetworkD3
}
\description{
After crawling organizational data, this is used for displaying an 
interactive graph with a force directed network of the units.
}
\examples{
\dontrun{
if(interactive()){
 abm_graph_divisions()
 }
}
}
\seealso{
\code{\link[networkD3]{forceNetwork}},\code{\link[networkD3]{JS}}
}
