% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_jcf}
\alias{abm_graph_jcf}
\title{Create graph over jcf by year}
\usage{
abm_graph_jcf(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_table4()}
}
\value{
a ggplot object
}
\description{
Create graph over jcf by year
}
