% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-plotly.R
\name{abm_graph_jcf_plotly}
\alias{abm_graph_jcf_plotly}
\title{Create graph over jcf by year, using plotly}
\usage{
abm_graph_jcf_plotly(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_table4()}
}
\value{
a plotly object
}
\description{
Create graph over jcf by year, using plotly
}
