% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_oa_lines}
\alias{abm_graph_oa_lines}
\title{Create line graph over share of publications by OA type and year}
\usage{
abm_graph_oa_lines(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_table6()}
}
\value{
a ggplot object
}
\description{
Create line graph over share of publications by OA type and year
}
