% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_oadata}
\alias{abm_graph_oadata}
\title{Create graph for Open Access data}
\usage{
abm_graph_oadata(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_oa_data()}
}
\value{
a pie chart
}
\description{
Create graph for Open Access data
}
