% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative-oa-graphs.R
\name{abm_graph_oadata_linegraphs}
\alias{abm_graph_oadata_linegraphs}
\title{Chart displaying the development of Open Access data publications over time}
\usage{
abm_graph_oadata_linegraphs(df, type = c("ggplot", "plotly", "ggiraph"))
}
\arguments{
\item{df}{a data frame at the format produced by abm_oa_data()}

\item{type}{the return type, one of "ggplot", "plotly" or "ggiraph"}
}
\value{
a ggplot object, ggiraph svg or plotly object as determined by type
}
\description{
This graph is an alternative or complement to the abm_graph_oadata_stackedarea graph and can return three types of results - a ggplot object, a plotly object or a ggiraph svg object
}
\details{
Rationale: \url{https://www.perceptualedge.com/articles/visual_business_intelligence/displays_for_combining_time-series_and_part-to-whole.pdf} and 
\url{https://everydayanalytics.ca/2014/08/stacked-area-graphs-are-not-your-friend.html}
}
