% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_oadata_pie3D}
\alias{abm_graph_oadata_pie3D}
\title{Create 3D pie chart for Open Access data}
\usage{
abm_graph_oadata_pie3D(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_oa_data()}
}
\value{
a 3D pie chart object
}
\description{
Create 3D pie chart for Open Access data
}
