% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative-oa-graphs.R
\name{abm_graph_oadata_piechart}
\alias{abm_graph_oadata_piechart}
\title{Pie chart for Open Access data}
\usage{
abm_graph_oadata_piechart(df, type = c("ggplot", "plotly", "ggiraph"))
}
\arguments{
\item{df}{a data frame at the format produced by abm_oa_data()}

\item{type}{the return type, one of "ggplot", "plotly" or "ggiraph"}
}
\value{
a ggplot object, ggiraph svg or plotly object as determined by type
}
\description{
This alternative pie chart can return three types of results - a ggplot object, a plotly object or a ggiraph svg object
}
