% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative-oa-graphs.R
\name{abm_graph_oadata_share}
\alias{abm_graph_oadata_share}
\title{Chart displaying shares of Open Access data publications}
\usage{
abm_graph_oadata_share(df, type = c("ggplot", "plotly", "ggiraph"))
}
\arguments{
\item{df}{a data frame at the format produced by abm_table6()}

\item{type}{the return type, one of "ggplot", "plotly" or "ggiraph"}
}
\value{
a ggplot object, ggiraph svg or plotly object as determined by type
}
\description{
This graph is an alternative or complement to the abm_graph_oadata_piechart graph and can return three types of results - a ggplot object, a plotly object or a ggiraph svg object
}
\details{
Rationale: <https://www.perceptualedge.com/articles/visual_business_intelligence/save_the_pies_for_dessert.pdf>
}
