% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_oadata_stackedarea}
\alias{abm_graph_oadata_stackedarea}
\title{Create stacked area graph for Open Access data}
\usage{
abm_graph_oadata_stackedarea(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_oa_data()}
}
\value{
a ggplot object
}
\description{
Create stacked area graph for Open Access data
}
