% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative-oa-graphs.R
\name{abm_graph_oadata_stackedarea_plotly}
\alias{abm_graph_oadata_stackedarea_plotly}
\title{Plotly chart wrapping the abm_graph_oadata_stacked graph}
\usage{
abm_graph_oadata_stackedarea_plotly(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_oa_data()}
}
\description{
Plotly chart wrapping the abm_graph_oadata_stacked graph
}
