% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_oadata_types_linechart}
\alias{abm_graph_oadata_types_linechart}
\title{Create graph over OA types by year}
\usage{
abm_graph_oadata_types_linechart(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_oa_data()}
}
\value{
a ggplot object
}
\description{
Create graph over OA types by year
}
