% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_scop_normcit}
\alias{abm_graph_scop_normcit}
\title{Create graph over Scopus FWCI by years}
\usage{
abm_graph_scop_normcit(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_table_scop_normcit()}
}
\value{
a ggplot object
}
\description{
Create graph over Scopus FWCI by years
}
