% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_scop_top10}
\alias{abm_graph_scop_top10}
\title{Create graph over Top 10\% publications by year, Scopus}
\usage{
abm_graph_scop_top10(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_table_scop_normcit()}
}
\value{
a ggplot object
}
\description{
Create graph over Top 10\% publications by year, Scopus
}
