% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_scop_top20}
\alias{abm_graph_scop_top20}
\title{Create graph over Top 20\% journals by year, Scopus}
\usage{
abm_graph_scop_top20(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_table4()}
}
\value{
a ggplot object
}
\description{
Create graph over Top 20\% journals by year, Scopus
}
