% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_sdg_mobile}
\alias{abm_graph_sdg_mobile}
\title{Create graph over SDGs, mobile version}
\usage{
abm_graph_sdg_mobile(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_sdg_table()}
}
\value{
a ggplot object
}
\description{
Create graph over SDGs, mobile version
}
