% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-plotly.R
\name{abm_graph_top10_plotly}
\alias{abm_graph_top10_plotly}
\title{Create graph over Cf by year, using plotly}
\usage{
abm_graph_top10_plotly(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_table3()}
}
\value{
a plotly object
}
\description{
Create graph over Cf by year, using plotly
}
