% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_graph_wos_coverage}
\alias{abm_graph_wos_coverage}
\title{Create graph over WoS coverage by year}
\usage{
abm_graph_wos_coverage(df)
}
\arguments{
\item{df}{a data frame at the format produced by abm_table1()}
}
\value{
a ggplot object
}
\description{
Create graph over WoS coverage by year
}
