% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_staffbased.R
\name{abm_oa_data_alt}
\alias{abm_oa_data_alt}
\title{Retrieve OA-data for ABM}
\usage{
abm_oa_data_alt(con = con_bib(), data)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{data}{dataset with publications as tibble}
}
\value{
tibble with OA-status of all publications from selected organizational units
}
\description{
Retrieve OA-data for ABM
}
