% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_plotlyfy}
\alias{abm_plotlyfy}
\title{Plotlyfy graph in ABM context}
\usage{
abm_plotlyfy(graphobj, tooltip, optionsbar = FALSE, showlegend = TRUE)
}
\arguments{
\item{graphobj}{a ggplot object}

\item{tooltip}{which paramenters to show as tooltips}

\item{optionsbar}{boolean, default false, indicating whether optionssbar should be shown}

\item{showlegend}{boolean, default true, indicating whether legend should be included}
}
\value{
a plotly object
}
\description{
Plotlyfy graph in ABM context
}
