% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_private_data}
\alias{abm_private_data}
\title{Private data from the Annual Bibliometric Monitoring project}
\usage{
abm_private_data(unit_code)
}
\arguments{
\item{unit_code}{the kthid for the researcher}
}
\value{
a list with two slots - "meta" for organizational unit metadata info 
  and "units" with a named list of results (set of 5 different tibbles for 
  the tables and also the publication list).
}
\description{
This returns an object which contains data for an individual researcher at KTH
}
\examples{
\dontrun{

# get all public data from the ABM
private <- abm_private_data('u1kzf1xh')

# get table 1 for the kthid
private \%>\% pluck('units', 'u1kzf1xh', 1)

# get publications for the kthid
private \%>\% pluck('units', 'u1kzf1xh', 'publications')

}   
}
