% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-bibliomatrix.R
\docType{data}
\name{abm_public_kth}
\alias{abm_public_kth}
\title{Public data with Annual Bibliometric Monitoring for KTH assembed in 2019}
\format{
An object of class \code{list} of length 4.
}
\usage{
abm_public_kth
}
\description{
A nested named list with two slots - a) "meta" with a tibble 
containing organizational unit information from KTH, including unit codes 
and b) "units" - a named list with a set of five tibbles per units
}
\details{
The data comes from a function call to \code{\link{abm_public_data}}
which requires database credentials being present in the .Renviron file. 
For information regarding how to access the data for various units at KTH,
please look at the examples for the \code{\link{abm_public_data}} function.
}
\keyword{datasets}
