% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_publications}
\alias{abm_publications}
\title{Retrieve publication list for ABM}
\usage{
abm_publications(
  data,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{data}{dataset with publications as tibble}

\item{analysis_start}{first publication year of analysis, default from abm_config()}

\item{analysis_stop}{last publication year of analysis, default from abm_config()}
}
\value{
tibble with publication list data for selected unit
}
\description{
Retrieve publication list for ABM
}
