% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_staffbased.R
\name{abm_publications_staffbased}
\alias{abm_publications_staffbased}
\title{Retrieve publication list for staffbased ABM}
\usage{
abm_publications_staffbased(
  con,
  unit_code,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{the kthids belonging to the analyzed unit}

\item{analysis_start}{first publication year of analysis, default 2012}

\item{analysis_stop}{last publication year of analysis, default 2018}
}
\value{
tibble with publication list data for selected unit
}
\description{
Retrieve publication list for staffbased ABM
}
