% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{abm_pubs_summary}
\alias{abm_pubs_summary}
\title{Summary for a list of publications}
\usage{
abm_pubs_summary(unit_slug)
}
\arguments{
\item{unit_slug}{identifier for the organizational unit}
}
\value{
tibble with summary
}
\description{
A list of publications belonging to for example an organizational unit
is summarized.
}
\examples{
\dontrun{
if(interactive()){
 abm_pubs_summary("j/jj/jjn")
 }
}
}
