% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{abm_report}
\alias{abm_report}
\title{A report used in ABM}
\usage{
abm_report(id, is_private)
}
\arguments{
\item{id}{the identifier for the report, either unit_code or kthid}

\item{is_private}{logical indicating whether the report is publicly visible or not}
}
\value{
a raw object with a BLOB containing a HTML report rendered using Rmarkdown.
}
\description{
A report used in ABM
}
\examples{
\dontrun{
if(interactive()){
 r1 <- abm_report(id = "KTH", is_private = FALSE)
 r2 <- abm_report("u1o2ujjd", is_private = TRUE)
 view_report(r2$data)
 }
}
}
\seealso{
\code{\link[kthapi]{kth_profile}}
}
