% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{abm_researchers}
\alias{abm_researchers}
\title{Researchers used in ABM}
\usage{
abm_researchers(unit_slug, con)
}
\arguments{
\item{unit_slug}{character identifier for a unit}

\item{con}{database connection, Default: con_bib()}
}
\value{
tibble with researchers
}
\description{
Researchers used in ABM
}
\examples{
\dontrun{
if(interactive()){
 abm_researchers()
 }
}
}
