% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_sdg_table}
\alias{abm_sdg_table}
\title{Create table over SDGs for the selected unit}
\usage{
abm_sdg_table(
  data,
  con,
  analysisId = abm_config()$analysis_id,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{data}{dataset with publications as tibble}

\item{con}{a database connection to BIBMON}

\item{analysisId}{id for the analysis, default from abm_config()}

\item{analysis_start}{first publication year of analysis, default from abm_config()}

\item{analysis_stop}{last publication year of analysis, default from abm_config()}
}
\value{
a tibble
}
\description{
Create table over SDGs for the selected unit
}
