% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{abm_slugs_departments}
\alias{abm_slugs_departments}
\title{Slugs or identifiers for departments at KTH used in ABM}
\usage{
abm_slugs_departments()
}
\value{
character vector of slugs
}
\description{
Slugs or identifiers for departments at KTH used in ABM
}
\examples{
\dontrun{
if(interactive()){
 slugs <- abm_slugs_departments()
 }
}
}
