% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{abm_slugs_institutions}
\alias{abm_slugs_institutions}
\title{abm_slugs_institutions}
\usage{
abm_slugs_institutions()
}
\value{
character vector of slugs
}
\description{
Slugs or identifiers for institutions at KTH used in ABM
}
\examples{
\dontrun{
if(interactive()){
 slugs <- abm_slugs_institutions()
 }
}
}
