% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_staffbased.R
\name{abm_staff_data}
\alias{abm_staff_data}
\title{Get publications from members of unit via kthids}
\usage{
abm_staff_data(con = con_bib(), kthids)
}
\arguments{
\item{con}{A database connection}

\item{kthids}{a list of KTH-ids to retrieve publications for}
}
\value{
tibble with all staff-based ABM data for selected organizational unit
}
\description{
Get publications from members of unit via kthids
}
