% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_staffbased.R
\name{abm_staff_list}
\alias{abm_staff_list}
\title{Get staff list from slug}
\usage{
abm_staff_list(
  con,
  unit_slug,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{con}{A database connection}

\item{unit_slug}{a string representing the KTH unit}

\item{analysis_start}{first publication year of analysis}

\item{analysis_stop}{last publication year of analysis}
}
\value{
tibble with staff list for selected organizational unit
}
\description{
Get staff list from slug
}
