% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bibmon.R
\name{abm_tab1}
\alias{abm_tab1}
\title{Retrieve data for the first ABM table}
\usage{
abm_tab1(con = con_bib())
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}
}
\value{
tibble with data for organizational units
}
\description{
Retrieve data for the first ABM table
}
