% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table1}
\alias{abm_table1}
\title{Retrieve Table 1 (Publications in DiVA) for ABM}
\usage{
abm_table1(
  con = con_bib(),
  unit_code,
  analysis_start = 2012,
  analysis_stop = 2018
)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{the code for the analyzed unit (KTH, a one letter school code, an integer department code or a KTH-id)}

\item{analysis_start}{first publication year of analysis, default 2012}

\item{analysis_stop}{last publication year of analysis, default 2018}
}
\value{
data frame with publications by type and year
}
\description{
Retrieve Table 1 (Publications in DiVA) for ABM
}
