% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table1_display}
\alias{abm_table1_display}
\title{Display Table 1 (Publications in DiVA) for web use}
\usage{
abm_table1_display(con = con_bib(), unit_code, pub_year)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{the code for the analyzed unit (KTH, a one letter school code, an integer department code or a KTH-id)}

\item{pub_year}{publication year(s) to analyze (optional, assuming master table holds only relevant years)}
}
\value{
data frame with publications by type and year
}
\description{
Display Table 1 (Publications in DiVA) for web use
}
