% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table1_full}
\alias{abm_table1_full}
\title{Retrieve Table 1 (Publications in DiVA) for ABM, full counts}
\usage{
abm_table1_full(
  data,
  con,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{data}{dataset with publications as tibble}

\item{con}{connection to db}

\item{analysis_start}{first publication year of analysis, default from abm_config()}

\item{analysis_stop}{last publication year of analysis, default from abm_config()}
}
\value{
data frame with publications by type and year
}
\description{
Retrieve Table 1 (Publications in DiVA) for ABM, full counts
}
