% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table2}
\alias{abm_table2}
\title{Retrieve Table 2 (Citations 3-year window) for ABM}
\usage{
abm_table2(
  data,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{data}{dataset with publications as tibble}

\item{analysis_start}{first publication year of analysis, default from abm_config()}

\item{analysis_stop}{last publication year of analysis, default from abm_config()}
}
\value{
tibble with citations statistics by year and total
}
\description{
Retrieve Table 2 (Citations 3-year window) for ABM
}
