% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_staffbased.R
\name{abm_table2_alt}
\alias{abm_table2_alt}
\title{Retrieve Table 2 (Citations 3-year window) for ABM}
\usage{
abm_table2_alt(
  con = con_bib(),
  data,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{data}{dataset with publications as tibble}

\item{analysis_start}{first publication year of analysis, default 2012}

\item{analysis_stop}{last publication year of analysis, default 2018}
}
\value{
tibble with citations statistics by year and total
}
\description{
Retrieve Table 2 (Citations 3-year window) for ABM
}
