% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table2_display}
\alias{abm_table2_display}
\title{Display Table 2 (Citations 3-year window) for ABM}
\usage{
abm_table2_display(con = con_bib(), unit_code, pub_year)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{the code for the analyzed unit (KTH, a one letter school code, an integer department code or a KTH-id)}

\item{pub_year}{publication year(s) to analyze (optional, assuming master table holds only relevant years)}
}
\value{
tibble with citations statistics by year and total
}
\description{
Display Table 2 (Citations 3-year window) for ABM
}
