% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table3}
\alias{abm_table3}
\title{Retrieve Table 3 (Field normalized citations) for ABM}
\usage{
abm_table3(
  con = con_bib(),
  unit_code,
  analysis_start = abm_defaults$start_year,
  analysis_stop = abm_defaults$stop_year
)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{the code for the analyzed unit (KTH, a one letter school code, an integer department code or a KTH-id)}

\item{analysis_start}{first publication year of analysis, default 2012}

\item{analysis_stop}{last publication year of analysis, default 2018}
}
\value{
tibble with field normalized citations and number/share of top10 publications by 3 year interval
}
\description{
Retrieve Table 3 (Field normalized citations) for ABM
}
