% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_staffbased.R
\name{abm_table3_alt}
\alias{abm_table3_alt}
\title{Retrieve Table 3 (Field normalized citations) for ABM}
\usage{
abm_table3_alt(
  con = con_bib(),
  data,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{data}{dataset with publications as tibble}

\item{analysis_start}{first publication year of analysis, default 2012}

\item{analysis_stop}{last publication year of analysis, default 2018}
}
\value{
tibble with field normalized citations and number/share of top10 publications by 3 year interval
}
\description{
Retrieve Table 3 (Field normalized citations) for ABM
}
