% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table3_display}
\alias{abm_table3_display}
\title{Display Table 3 (Field normalized citations) for ABM}
\usage{
abm_table3_display(con = con_bib(), unit_code, pub_year)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{the code for the analyzed unit (KTH, a one letter school code, an integer department code or a KTH-id)}

\item{pub_year}{publication year(s) to analyze (optional, assuming master table holds only relevant years)}
}
\value{
tibble with field normalized citations and number/share of top10 publications by 3 year interval
}
\description{
Display Table 3 (Field normalized citations) for ABM
}
