% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table4}
\alias{abm_table4}
\title{Retrieve Table 4 (Journal impact) for ABM}
\usage{
abm_table4(
  con = con_bib(),
  unit_code,
  analysis_start = abm_defaults$start_year,
  analysis_stop = abm_defaults$stop_year
)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{the code for the analyzed unit (KTH, a one letter school code, an integer department code or a KTH-id)}

\item{analysis_start}{first publication year of analysis, default 2012}

\item{analysis_stop}{last publication year of analysis, default 2018}
}
\value{
tibble with field normalized journal citation score and number/share of publications in top20 journals
}
\description{
Retrieve Table 4 (Journal impact) for ABM
}
