% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table5}
\alias{abm_table5}
\title{Retrieve Table 5 (Co-publishing) for ABM}
\usage{
abm_table5(
  con = con_bib(),
  unit_code,
  analysis_start = abm_defaults$start_year,
  analysis_stop = abm_defaults$stop_year
)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{the code for the analyzed unit (KTH, a one letter school code, an integer department code or a KTH-id)}

\item{analysis_start}{first publication year of analysis, default 2012}

\item{analysis_stop}{last publication year of analysis, default 2018}
}
\value{
tibble with number/share of international copublications and copublications with Swedish non-university organizations
}
\description{
Retrieve Table 5 (Co-publishing) for ABM
}
