% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table5_display}
\alias{abm_table5_display}
\title{Display Table 5 (Co-publishing) for ABM}
\usage{
abm_table5_display(con = con_bib(), unit_code, pub_year)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{the code for the analyzed unit (KTH, a one letter school code, an integer department code or a KTH-id)}

\item{pub_year}{publication year(s) to analyze (optional, assuming master table holds only relevant years)}
}
\value{
tibble with number/share of international copublications and copublications with Swedish non-university organizations
}
\description{
Display Table 5 (Co-publishing) for ABM
}
