% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table6}
\alias{abm_table6}
\title{Retrieve Table 6 (OA data) for ABM}
\usage{
abm_table6(con = con_bib(), unit_code)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{for filtering on one or more unit code(s), which can be KTH, a one letter school code, an integer department code or a KTH-id (optional)}
}
\value{
tibble with OA-status of all publications from selected organizational units
}
\description{
Retrieve Table 6 (OA data) for ABM
}
