% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table_scop_normcit}
\alias{abm_table_scop_normcit}
\title{Retrieve field normalized citations table (Scopus) for ABM}
\usage{
abm_table_scop_normcit(
  data,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{data}{dataset with publications as tibble}

\item{analysis_start}{first publication year of analysis, default from abm_config()}

\item{analysis_stop}{last publication year of analysis, default from abm_config()}
}
\value{
tibble with field normalized citations and number/share of top10 publications by 3 year interval
}
\description{
Retrieve field normalized citations table (Scopus) for ABM
}
