% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table_scop_snip}
\alias{abm_table_scop_snip}
\title{Retrieve Journal impact table (Scopus) for ABM}
\usage{
abm_table_scop_snip(
  data,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year,
  intervals = FALSE
)
}
\arguments{
\item{data}{dataset with publications as tibble}

\item{analysis_start}{first publication year of analysis, default from abm_config()}

\item{analysis_stop}{last publication year of analysis, default from abm_config()}

\item{intervals}{set to TRUE to use 3 year intervals, default FALSE}
}
\value{
tibble with field normalized journal citation score and number/share of publications in top20 journals
}
\description{
Retrieve Journal impact table (Scopus) for ABM
}
