% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_bullet_citations}
\alias{abm_ui_bullet_citations}
\title{Bullet graph for citations in ABM}
\usage{
abm_ui_bullet_citations(df)
}
\arguments{
\item{df}{data frame with citation stats by 3year interval}
}
\description{
Bullet graph for citations in ABM
}
