% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_bullet_journal}
\alias{abm_ui_bullet_journal}
\title{Bullet graph for journal citations in ABM}
\usage{
abm_ui_bullet_journal(df)
}
\arguments{
\item{df}{data frame with journal citation stats by year}
}
\description{
Bullet graph for journal citations in ABM
}
