% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_button_altmetric}
\alias{abm_ui_button_altmetric}
\title{Link to Altmetric Explorer for an organizational unit at KTH}
\usage{
abm_ui_button_altmetric(altmetric_count, altmetric_href, unit_label)
}
\arguments{
\item{altmetric_count}{the count to include in a tooltip}

\item{altmetric_href}{the target of the link}

\item{unit_label}{the label for the organizational unit}
}
\description{
Link to Altmetric Explorer for an organizational unit at KTH
}
