% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_button_publist}
\alias{abm_ui_button_publist}
\title{Link to Publication List download for KTH}
\usage{
abm_ui_button_publist(is_loggedin, unit_label, unit_code, unit_file_label)
}
\arguments{
\item{is_loggedin}{boolean parameter indicating logged in status}

\item{unit_label}{the label for the organizational unit}

\item{unit_code}{the unit code for the organizational unit}

\item{unit_file_label}{the label for the download}
}
\description{
Link to Publication List download for KTH
}
