% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_staffbased.R
\name{abm_ui_button_publist_authbased}
\alias{abm_ui_button_publist_authbased}
\title{Link to Publication List download for authorbased ABM}
\usage{
abm_ui_button_publist_authbased(is_loggedin, data, unit_label, unit_file_label)
}
\arguments{
\item{is_loggedin}{boolean parameter indicating logged in status}

\item{unit_label}{the label for the organizational unit}

\item{unit_file_label}{the label for the download}

\item{unit_code}{the unit code for the organizational unit}
}
\description{
Link to Publication List download for authorbased ABM
}
