% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_datatable_cf}
\alias{abm_ui_datatable_cf}
\title{Datatable for 3 year citations (cf)}
\usage{
abm_ui_datatable_cf(df_cf, unit_file_label, unit_title)
}
\arguments{
\item{df_cf}{data frame with DiVA publication data in a specific format}

\item{unit_file_label}{the filename presented when users make use of the download button}

\item{unit_title}{the label presented when users make use of the download button}
}
\description{
Datatable for 3 year citations (cf)
}
