% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_datatable_copub_countries}
\alias{abm_ui_datatable_copub_countries}
\title{Datatable for co-publication countries (WoS)}
\usage{
abm_ui_datatable_copub_countries(
  df_copub_countries,
  unit_file_label,
  unit_title
)
}
\arguments{
\item{df_copub_countries}{data frame with co-publication countries and number of publications}

\item{unit_file_label}{the filename presented when users make use of the download button}

\item{unit_title}{the label presented when users make use of the download button}
}
\description{
Datatable for co-publication countries (WoS)
}
