% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_datatable_diva_full}
\alias{abm_ui_datatable_diva_full}
\title{Datatable for DiVA publications, full counts}
\usage{
abm_ui_datatable_diva_full(df_diva_full, unit_file_label, unit_title)
}
\arguments{
\item{df_diva_full}{data frame with DiVA publication data in a specific format}

\item{unit_file_label}{the filename presented when users make use of the download button}

\item{unit_title}{the label presented when users make use of the download button}
}
\description{
Datatable for DiVA publications, full counts
}
