% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_datatable_jcf}
\alias{abm_ui_datatable_jcf}
\title{Datatable for journal impact (jcf)}
\usage{
abm_ui_datatable_jcf(df_jcf, unit_file_label, unit_title)
}
\arguments{
\item{df_jcf}{data frame with DiVA publication data in a specific format}

\item{unit_file_label}{the filename presented when users make use of the download button}

\item{unit_title}{the label presented when users make use of the download button}
}
\description{
Datatable for journal impact (jcf)
}
