% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_datatable_publications}
\alias{abm_ui_datatable_publications}
\title{Datatable for publication list}
\usage{
abm_ui_datatable_publications(data, unit_title, unit_file_label)
}
\arguments{
\item{data}{data frame with list of publications}

\item{unit_title}{the label presented when users make use of the download button}

\item{unit_file_label}{the filename presented when users make use of the download button}
}
\description{
Datatable for publication list
}
