% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_datatable_researchers}
\alias{abm_ui_datatable_researchers}
\title{Datatable for researchers}
\usage{
abm_ui_datatable_researchers(data, unit_file_label, unit_title)
}
\arguments{
\item{data}{data frame with researchers in a specific format (needs name, title)}

\item{unit_file_label}{the filename presented when users make use of the download button}

\item{unit_title}{the label presented when users make use of the download button}
}
\description{
Datatable for researchers
}
