% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_datatable_sdg_year}
\alias{abm_ui_datatable_sdg_year}
\title{Datatable for SDG publications by year}
\usage{
abm_ui_datatable_sdg_year(df_sdg_year, unit_file_label, unit_title)
}
\arguments{
\item{df_sdg_year}{data frame with data on SDG pubs}

\item{unit_file_label}{the filename presented when users make use of the download button}

\item{unit_title}{the label presented when users make use of the download button}
}
\description{
Datatable for SDG publications by year
}
