% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_cf}
\alias{abm_ui_kable_cf}
\title{HTML table for 3 year citations (cf)}
\usage{
abm_ui_kable_cf(df_cf)
}
\arguments{
\item{df_cf}{data frame with DiVA publication data in a specific format}
}
\description{
HTML table for 3 year citations (cf)
}
