% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_city3y}
\alias{abm_ui_kable_city3y}
\title{HTML table for 3 year citations}
\usage{
abm_ui_kable_city3y(df_city3y)
}
\arguments{
\item{df_city3y}{data frame with DiVA publication data in a specific format}
}
\description{
HTML table for 3 year citations
}
