% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_copub}
\alias{abm_ui_kable_copub}
\title{HTML table for co-publication (WoS)}
\usage{
abm_ui_kable_copub(df_copub)
}
\arguments{
\item{df_copub}{data frame with DiVA publication data in a specific format}
}
\description{
HTML table for co-publication (WoS)
}
