% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_copub_countries}
\alias{abm_ui_kable_copub_countries}
\title{HTML table for co-publication countries (WoS)}
\usage{
abm_ui_kable_copub_countries(df_copub_countries)
}
\arguments{
\item{df_copub_countries}{data frame with co-publication data in a specific format}
}
\description{
HTML table for co-publication countries (WoS)
}
