% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_copub_orgs}
\alias{abm_ui_kable_copub_orgs}
\title{HTML table for co-publication organizations (WoS)}
\usage{
abm_ui_kable_copub_orgs(df_copub_orgs)
}
\arguments{
\item{df_copub_orgs}{data frame with co-publication data in a specific format}
}
\description{
HTML table for co-publication organizations (WoS)
}
