% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_diva}
\alias{abm_ui_kable_diva}
\title{HTML table for DiVA publications}
\usage{
abm_ui_kable_diva(df_diva)
}
\arguments{
\item{df_diva}{data frame with DiVA publication data in a specific format}
}
\description{
HTML table for DiVA publications
}
