% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_diva_full}
\alias{abm_ui_kable_diva_full}
\title{HTML table for DiVA publications, full counts}
\usage{
abm_ui_kable_diva_full(df_diva_full)
}
\arguments{
\item{df_diva_full}{data frame with DiVA publication data in a specific format}
}
\description{
HTML table for DiVA publications, full counts
}
