% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_jcf}
\alias{abm_ui_kable_jcf}
\title{HTML table for journal impact (jcf)}
\usage{
abm_ui_kable_jcf(df_jcf)
}
\arguments{
\item{df_jcf}{data frame with DiVA publication data in a specific format}
}
\description{
HTML table for journal impact (jcf)
}
