% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_oa}
\alias{abm_ui_kable_oa}
\title{HTML table for open access publications}
\usage{
abm_ui_kable_oa(df_oa)
}
\arguments{
\item{df_oa}{data frame with DiVA publication data in a specific format}
}
\description{
HTML table for open access publications
}
