% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_scop_cit}
\alias{abm_ui_kable_scop_cit}
\title{HTML table for Scopus citations}
\usage{
abm_ui_kable_scop_cit(df_scop_cit)
}
\arguments{
\item{df_scop_cit}{data frame with DiVA publication data in a specific format}
}
\description{
HTML table for Scopus citations
}
