% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_scop_copub}
\alias{abm_ui_kable_scop_copub}
\title{HTML table for co-publication (Scopus)}
\usage{
abm_ui_kable_scop_copub(df_scop_copub)
}
\arguments{
\item{df_scop_copub}{data frame with DiVA publication data in a specific format}
}
\description{
HTML table for co-publication (Scopus)
}
