% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_scop_snip}
\alias{abm_ui_kable_scop_snip}
\title{HTML table for Scopus journal impact (SNIP)}
\usage{
abm_ui_kable_scop_snip(df_scop_snip)
}
\arguments{
\item{df_scop_snip}{data frame with DiVA publication data in a specific format}
}
\description{
HTML table for Scopus journal impact (SNIP)
}
