% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_kable_sdg_table}
\alias{abm_ui_kable_sdg_table}
\title{HTML table for SDG publicaitons by goal}
\usage{
abm_ui_kable_sdg_table(df_sdg_year)
}
\arguments{
\item{df_sdg_year}{data frame with data on SDG pubs}
}
\description{
HTML table for SDG publicaitons by goal
}
