% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_note}
\alias{abm_ui_note}
\title{A note to keep in mind when interpreting results}
\usage{
abm_ui_note(
  data,
  df_coverage,
  unit_level,
  is_fractional = FALSE,
  is_wos = TRUE
)
}
\arguments{
\item{data}{data frame with DiVA publication data in a specific format}

\item{df_coverage}{data frame with coverage data}

\item{unit_level}{an integer indicating organizational aggregation level}

\item{is_fractional}{logical indicating if fractional values, by default FALSE}

\item{is_wos}{logical indicating if using WoS data, by default TRUE}
}
\description{
A note to keep in mind when interpreting results
}
