% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_summary_pubs}
\alias{abm_ui_summary_pubs}
\title{Summary valuebox for publications}
\usage{
abm_ui_summary_pubs(df_diva, lastyear, start_year, stop_year, unit_label)
}
\arguments{
\item{df_diva}{data frame with DiVA publication data in a specific format}

\item{lastyear}{previous year}

\item{start_year}{first year of total interval of years}

\item{stop_year}{last year of total interval of years}

\item{unit_label}{the unit label to use in case of no data}
}
\description{
Summary valuebox for publications
}
