% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_valuebox_coverage}
\alias{abm_ui_valuebox_coverage}
\title{Value box for coverage in ABM}
\usage{
abm_ui_valuebox_coverage(df, vbcolor, db = c("wos", "scopus"))
}
\arguments{
\item{df}{data frame with coverage data}

\item{vbcolor}{color of valueBox}

\item{db}{database for which coverage will be shown ("wos" or "scopus")}
}
\description{
Value box for coverage in ABM
}
