% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_waffle_copub}
\alias{abm_ui_waffle_copub}
\title{Waffle graph for co-publications in ABM}
\usage{
abm_ui_waffle_copub(df)
}
\arguments{
\item{df}{data frame with co-publications stats by year}
}
\description{
Waffle graph for co-publications in ABM
}
