% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{abm_unit_pubs}
\alias{abm_unit_pubs}
\title{Publications for ABM researchers belonging to a unit}
\usage{
abm_unit_pubs(unit_slug)
}
\arguments{
\item{unit_slug}{identifier for organizational unit}
}
\value{
tibble with publications
}
\description{
Publications corresponding to researchers associated with a unit slug are 
returned.
}
\examples{
\dontrun{
if(interactive()){
 abm_unit_pubs("j/jj/jjn")
 }
}
}
