% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{abm_unit_stats}
\alias{abm_unit_stats}
\title{Publication summary stats for ABM divisions and institutions and schools}
\usage{
abm_unit_stats()
}
\value{
tibble with summary data
}
\description{
Publication summary stats for ABM divisions and institutions and schools
}
\examples{
\dontrun{
if(interactive()){
 stats <- abm_unit_stats() 
 # upload results to database
 db_upsert_table("unit_stats", stats)
 }
}
}
