% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bibmon.R
\name{abm_units}
\alias{abm_units}
\title{Organizational units at KTH}
\usage{
abm_units(con = con_bib())
}
\arguments{
\item{con}{connection to db, default is to use mssql db con}
}
\value{
tibble
}
\description{
Organizational units at KTH
}
