% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_waffle_copub_int}
\alias{abm_waffle_copub_int}
\title{Create waffle share over share of international copublications}
\usage{
abm_waffle_copub_int(df, col = c(as.character(palette_kth(1)), "#F6F6F6"))
}
\arguments{
\item{df}{a data frame at the format produced by abm_table5()}
}
\value{
a ggplot object
}
\description{
Create waffle share over share of international copublications
}
