% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_waffle_pct}
\alias{abm_waffle_pct}
\title{Create waffle chart (5 rows, 20 columns) for any single percentage}
\usage{
abm_waffle_pct(pct, label, col)
}
\arguments{
\item{pct}{a percentage expressed as a decimal number 0 <= pct <= 1}

\item{label}{a title for the chart, displayed above the waffle (optional)}

\item{col}{a vector with colors for filling (optional)}
}
\value{
a ggplot object
}
\description{
Create waffle chart (5 rows, 20 columns) for any single percentage
}
