% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-plotly.R
\name{abm_waffle_pct_plotly}
\alias{abm_waffle_pct_plotly}
\title{Create alternate waffle chart (with plotly)}
\usage{
abm_waffle_pct_plotly(pct, col, label = NULL)
}
\arguments{
\item{pct}{a percentage expressed as a decimal number 0 <= pct <= 1}

\item{col}{a vector with colors for filling (optional)}

\item{label}{a title for the chart, displayed above the waffle (optional)}
}
\value{
a plotly object

a plotly object
}
\description{
Create alternate waffle chart (with plotly)
}
