% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_woscoverage}
\alias{abm_woscoverage}
\title{Retrieve WoS coverage for peer reviewed DiVA publication types}
\usage{
abm_woscoverage(
  con,
  unit_code,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{the code for the analyzed unit (KTH, a one letter school code, an integer department code or a KTH-id)}

\item{analysis_start}{first publication year of analysis, if not given abm_config() is used}

\item{analysis_stop}{last publication year of analysis, if not given abm_config() is used}
}
\value{
tibble with fractionalized and full counted WoS coverage by year and publication type
}
\description{
Retrieve WoS coverage for peer reviewed DiVA publication types
}
