% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{ad_accountname}
\alias{ad_accountname}
\title{LDAP account name from a given KTH id}
\usage{
ad_accountname(kthid)
}
\arguments{
\item{kthid}{the employee identifier}
}
\description{
LDAP account name from a given KTH id
}
