% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{ad_displayname}
\alias{ad_displayname}
\title{KTH displayname from user account identifier}
\usage{
ad_displayname(kthid)
}
\arguments{
\item{kthid}{the employee identifier}
}
\description{
KTH displayname from user account identifier
}
