% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthapi.R
\name{ad_displayname2}
\alias{ad_displayname2}
\title{KTH displayname from user account identifier}
\usage{
ad_displayname2(kthid)
}
\arguments{
\item{kthid}{the employee identifier}
}
\description{
KTH displayname from user account identifier
}
