% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{ad_kthid}
\alias{ad_kthid}
\title{KTH user account identifier (kthid) from LDAP account name}
\usage{
ad_kthid(accountname)
}
\arguments{
\item{accountname}{the LDAP accountname}
}
\description{
KTH user account identifier (kthid) from LDAP account name
}
