% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthapi.R
\name{ad_kthid2}
\alias{ad_kthid2}
\title{KTH user account identifier (kthid) from account name}
\usage{
ad_kthid2(accountname)
}
\arguments{
\item{accountname}{the LDAP accountname}
}
\description{
KTH user account identifier (kthid) from account name
}
