% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{ad_memberof}
\alias{ad_memberof}
\title{List member of properties from LDAP search}
\usage{
ad_memberof(search_results)
}
\arguments{
\item{search_results}{a tibble with search results from the `ad_search()` fcn}
}
\value{
tibble with memberof LDIF values
}
\description{
List member of properties from LDAP search
}
