% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{ad_search}
\alias{ad_search}
\title{Search Active Directory at KTH using kthid or username}
\usage{
ad_search(search_term, search_type = c("kthid", "accountname"))
}
\arguments{
\item{search_term}{the KTH id or account name to use in the search query}

\item{search_type}{one of "kthid" or "accountname" for the type of search to make}
}
\value{
tibble with key value pairs
}
\description{
This function uses curl to query the KTH Active Directory. It requires 
environment variables to be set in .Renviron, specifically
LDAP_USER, LDAP_PASS, LDAP_HOST and LDAP_BASE for the service 
account used for the queries.
}
\details{
Under the hood, this function issues a curl request similar to this one:
curl --ntlm -user $CREDS \\
ldaps://ldap.foo.com/DC=ads,DC=foo,DC=com?memberOf?sub?(&(sAMAccountName=$USER)(memberOf=CN=$GROUP,OU=Distribution,OU=Groups,DC=ads,DC=foo,DC=com))"
}
\examples{
ad_search("markussk", "accountname")
}
