% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-bibliomatrix.R
\name{bibliomatrix}
\alias{bibliomatrix}
\alias{package-bibliomatrix}
\title{bibliomatrix}
\description{
The goal of bibliomatrix is to provide a common set of functionality for 
the Annual Bibliometric Monitoring efforts from the Bibliometrics team at 
the KTH Library.
}
\keyword{package}
