% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{coveragetext}
\alias{coveragetext}
\title{Generate text for Web of Science coverage warnings}
\usage{
coveragetext(woscoverage)
}
\arguments{
\item{woscoverage}{Web of Science coverage (between 0 and 1)}
}
\value{
a string with an information text for the correct interval
}
\description{
Generate text for Web of Science coverage warnings
}
