% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_sqlite_location}
\alias{db_sqlite_location}
\title{Location of sqlite3 db file for BIBMON data}
\usage{
db_sqlite_location()
}
\value{
character string representing on disk location for db file
}
\description{
Location of sqlite3 db file for BIBMON data
}
