% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{db_upload_crawl}
\alias{db_upload_crawl}
\title{Crawl and persist organizational data from KTH API}
\usage{
db_upload_crawl(con = con_bib(), crawl = FALSE)
}
\arguments{
\item{con}{connection to database, Default: con_bib()}

\item{crawl}{logical flag confirming intention to perform crawl, Default: FALSE}
}
\value{
invisible TRUE on success
}
\description{
This function can be used to trigger a crawl for organizational data provided
by the KTH APIs. Data will be persisted using the provided database connection
}
\details{
The parameter crawl must be set to TRUE to run this function
}
\examples{
\dontrun{
if(interactive()){
 db_upload_crawl()
 }
}
}
