% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{db_upsert_table}
\alias{db_upsert_table}
\title{Create or overwrite table at connection source}
\usage{
db_upsert_table(tbl, df, con = con_bib())
}
\arguments{
\item{tbl}{table name}

\item{df}{data frame or tibble with the data}

\item{con}{database connection, Default: con_bib()}
}
\value{
invisibly TRUE on success
}
\description{
Utility that can be used to sync data against a connection.
}
\details{
not all backend drivers support overwrite and append, therefore
param overwrite for dbWriteTable (see docs) is not used
}
\examples{
\dontrun{
if(interactive()){
 db_upsert_table("divisions", abm_divisions(), con = con_bib(type = "mssql"))
 }
}
}
\seealso{
\code{\link[DBI]{dbExistsTable}},\code{\link[DBI]{dbRemoveTable}},\code{\link[DBI]{dbWriteTable}}
}
