% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthapi.R
\name{displayname_from_kthid}
\alias{displayname_from_kthid}
\title{KTH displayname from user account identifier}
\usage{
displayname_from_kthid(kthid)
}
\arguments{
\item{kthid}{the employee identifier}
}
\description{
KTH displayname from user account identifier
}
