% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{get_analysis_date}
\alias{get_analysis_date}
\title{Get date of indicator extraction for ABM}
\usage{
get_analysis_date(con)
}
\arguments{
\item{con}{connection to db - if no connection is given, use abm_public_kth data}
}
\value{
date
}
\description{
Get date of indicator extraction for ABM
}
