% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{get_indic_descriptions}
\alias{get_indic_descriptions}
\title{Get displayname and descriptions for indicators}
\usage{
get_indic_descriptions(con)
}
\arguments{
\item{con}{connection to db - if no connection is given, use abm_public_kth data}
}
\value{
tibble with names and descriptions
}
\description{
Get displayname and descriptions for indicators
}
