% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{get_pubtype_order}
\alias{get_pubtype_order}
\title{Get order of publication type for ABM table 1}
\usage{
get_pubtype_order(con)
}
\arguments{
\item{con}{connection to db - if no connection is given, attempt to use abm_public_kth data}
}
\value{
tibble with pt_ordning and diva_publiation_type
}
\description{
Get order of publication type for ABM table 1
}
