% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{kth_catalog_crawl}
\alias{kth_catalog_crawl}
\title{kth_catalog_crawl}
\usage{
kth_catalog_crawl(slug)
}
\arguments{
\item{slug}{a string with the slug, for example "j/jj"}
}
\value{
a tibble encoding descendants as a hierarchy (using parent_id, child_id tuples)
}
\description{
Given an organizational "slug", a depth first traversal is made
enumerating organizational units (descendants)
}
\details{
this fcn should be moved to the kthapi package... please avoid using this fcn too often as it descends recursively and
generates several API calls; instead please cache the results and try to avoid hammering the API
}
\examples{
\dontrun{
if(interactive()){
 kth_catalog_crawl("j/jj")
 }
}
}
\seealso{
\code{\link[kthapi]{kth_catalog}}
 \code{\link[purrr]{map}}
 \code{\link[dplyr]{bind}}
}
