% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{kth_divisions_crawl}
\alias{kth_divisions_crawl}
\title{Crawl kthapi for divisions}
\usage{
kth_divisions_crawl(
  include = abm_slugs_departments(),
  exclude = NULL,
  quiet = FALSE
)
}
\arguments{
\item{include}{character vector of slugs to include, Default: abm_slugs_institutions()}

\item{exclude}{character vector of slugs to exclude, Default: NULL}

\item{quiet}{logical to indicate logging, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function crawls the KTH Directory API for divisions, by default those used in ABM
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 kth_divisions_crawl()
 }
}
}
\seealso{
\code{\link[progress]{progress_bar}}
 \code{\link[purrr]{safely}},\code{\link[purrr]{map}}
}
