% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{kth_theme}
\alias{kth_theme}
\title{KTH plot theme for ggplot}
\usage{
kth_theme(base_size = 12, base_font = "Arial")
}
\arguments{
\item{base_size}{the basic font size used in graphs}

\item{base_font}{the default font used in graphs}
}
\value{
a ggplot theme
}
\description{
KTH plot theme for ggplot
}
