% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthapi.R
\name{kthid_from_accountname}
\alias{kthid_from_accountname}
\title{KTH user account identifier (kthid) from account name}
\usage{
kthid_from_accountname(accountname)
}
\arguments{
\item{accountname}{the LDAP accountname}
}
\description{
KTH user account identifier (kthid) from account name
}
