% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{mean_indicator_units}
\alias{mean_indicator_units}
\title{Calculate average citation indicators across years}
\usage{
mean_indicator_units(
  con,
  starty,
  stopy,
  analysis_level = 2,
  analysis_version_id
)
}
\arguments{
\item{con}{connection to db}

\item{starty}{first publication year of analysis}

\item{stopy}{last publication year of analysis}

\item{analysis_level}{organization analysis level. Default is 2 (department).}

\item{analysis_version_id}{the analysis_id id to be used from masterfile}
}
\value{
tibble average citation indicators along with dept name, school name, along with full and fractional publ. counts.
}
\description{
This function calculates average indicator values (jcf, cf etc) across a set of years per department, using fractional counting.
Data is based on masterfile, using a specific analysis_id (i.e. data version nr.)
This three-year average of jcf is used as an performance indicator at KTH.
}
