% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{mean_jcf}
\alias{mean_jcf}
\title{Calculate average jcf across years}
\usage{
mean_jcf(con, starty, stopy, analysis_level = 2)
}
\arguments{
\item{con}{connection to db}

\item{starty}{first publication year of analysis}

\item{stopy}{last publication year of analysis}

\item{analysis_level}{organization analysis level. Default is 2 (department).}
}
\value{
tibble average jcf along with dept name, school name, along with full and fractional publ. counts.
}
\description{
This function calculates average jcf across a set of years per department.
This three-year average is used as an performance indicator at KTH.
}
