% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{mean_jcf_units}
\alias{mean_jcf_units}
\title{Calculate average jcf across years}
\usage{
mean_jcf_units(con, starty, stopy, analysis_level = 2, analysis_version_id)
}
\arguments{
\item{con}{connection to db}

\item{starty}{first publication year of analysis}

\item{stopy}{last publication year of analysis}

\item{analysis_level}{organization analysis level. Default is 2 (department).}

\item{analysis_version_id}{the analysis_id id to be used from masterfile}
}
\value{
tibble average jcf along with dept name, school name, along with full and fractional publ. counts.
}
\description{
This function calculates average jcf across a set of years per department, using fractional counting.
Data is based on masterfile, using a specific analysis_id (i.e. data version nr.)
This three-year average is used as an performance indicator at KTH.
}
