% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formatting.R
\name{nowrap_html}
\alias{nowrap_html}
\title{String formatting utility to avoid wrapping of hyphens inside compound words
while retaining search in HTML content using Ctrl-F.}
\usage{
nowrap_html(x)
}
\arguments{
\item{x}{the string to modify}
}
\description{
For example, if "co-publications" should be possible to find on a HTML page
using Ctrl-F, a non-breaking hyphen (longer dash) HTML entity might not be ideal.
}
\details{
Wrapping the string "co-publication" in a span with the CSS style 'white-space:nowrap;'
might be more convenient for the end user.
}
\examples{
nowrap_html("Swedish non-university co-publications")
}
