% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formatting.R
\name{nowrap_hyphen_sub}
\alias{nowrap_hyphen_sub}
\title{String formatting utility to substitute soft hyphens inside compound words
with non-breaking hard "long" hyphens using HTML character entity encoding}
\usage{
nowrap_hyphen_sub(x)
}
\arguments{
\item{x}{the string to modify}
}
\description{
For example, "co-publications" becomes "co&#8209;publications"
}
\examples{
nowrap_hyphen_sub("Swedish non-university co-publications")
}
