% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{palette_kth}
\alias{palette_kth}
\title{KTH color palette with 5 qualitative colors}
\usage{
palette_kth(n = 5, name = "KTH")
}
\arguments{
\item{n}{number of colors to use (1..5), default is to return all five}

\item{name}{name of palette}
}
\value{
named vector with colors and hex codes
}
\description{
This palette is taken from https://intra.kth.se/en/administration/kommunikation/grafiskprofil/profilfarger-1.845077
}
\details{
This function aligns its signature with RColorBrewer::color.pal()
}
\examples{
palette_kth(1)  # return primary color in the KTH palette
}
