% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{palette_kth_info}
\alias{palette_kth_info}
\title{KTH color palette information}
\usage{
palette_kth_info()
}
\value{
data frame with information about the available palette(s)
}
\description{
This function aligns its signature with RColorBrewer::color.pal.info()
}
