% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{pool_bib}
\alias{pool_bib}
\title{Connection pool to Bibliometrics data source for KTH}
\usage{
pool_bib(source_type = c("sqlite", "mssql"))
}
\arguments{
\item{source_type}{one of "sqlite" or "mssql" with "sqlite" being default}
}
\value{
database connection
}
\description{
This function returns a db connection to one of two possible pre-configured
data sources containing Bibliometrics data
}
