% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{prerender}
\alias{prerender}
\title{Prerender dashboards for public data}
\usage{
prerender(refresh = FALSE)
}
\arguments{
\item{refresh}{logical to indicate whether the cache should be refreshed}
}
\description{
Prerender dashboards for public data
}
