% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{prerender_cache_location}
\alias{prerender_cache_location}
\title{Location for prerender cache with dashboard HTML content}
\usage{
prerender_cache_location()
}
\description{
Location for prerender cache with dashboard HTML content
}
