% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{prerender_cache_sync}
\alias{prerender_cache_sync}
\title{Update shiny apps www/cache with prerenderd 
dashboard HTML content}
\usage{
prerender_cache_sync(overwrite = FALSE)
}
\arguments{
\item{overwrite}{logical boolean default false}
}
\description{
Update shiny apps www/cache with prerenderd 
dashboard HTML content
}
