% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{run_api}
\alias{run_api}
\title{Running this function launches a Plumber API presenting
resources bundled in this package}
\usage{
run_api(port = 8080, host = "0.0.0.0")
}
\arguments{
\item{port}{the port to run on, default is 8080}

\item{host}{the interface to bind to, default is all 0.0.0.0}
}
\description{
Running this function launches a Plumber API presenting
resources bundled in this package
}
