% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{sliding_intervals}
\alias{sliding_intervals}
\title{Create integer intervals useful for e.g. sliding means}
\usage{
sliding_intervals(first, last, width)
}
\arguments{
\item{first}{the smallest integer in the range to use}

\item{last}{the largest integer in the range to use}

\item{width}{the desired width of intervals}
}
\value{
data frame with label for each interval and one row for each year
}
\description{
Create integer intervals useful for e.g. sliding means
}
