% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status_db}
\alias{status_db}
\title{Status message related to db availability}
\usage{
status_db()
}
\value{
list with two slots for status message and status
}
\description{
Status message related to db availability
}
