% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status_kthapi}
\alias{status_kthapi}
\title{Status message related to KTH API availability}
\usage{
status_kthapi()
}
\value{
list with two slots for status message and status
}
\description{
Status message related to KTH API availability
}
