% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status_renviron}
\alias{status_renviron}
\title{Status message related to R environment file availability}
\usage{
status_renviron()
}
\value{
list with two slots for status message and status
}
\description{
Status message related to R environment file availability
}
