% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{timegraph_header_legend}
\alias{timegraph_header_legend}
\title{Graph to subheader of timegraph}
\usage{
timegraph_header_legend(colors_vb = ktheme::palette_kth_neo(17))
}
\arguments{
\item{colors_vb}{a color set to use}
}
\description{
Graph to subheader of timegraph
}
