% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{unit_info}
\alias{unit_info}
\title{Retrieve information about ABM units (level 0-2)}
\usage{
unit_info(con = con_bib(), unit, level, parent)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit}{unit code(s) for selection}

\item{level}{organizational level(s) for selection (0 = KTH, 1 = School, 2 = Department)}

\item{parent}{selects organizations with specific parent(s)}
}
\value{
tibble with information about selected units
}
\description{
Retrieve information about ABM units (level 0-2)
}
