% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_staffbased.R
\name{unit_staff}
\alias{unit_staff}
\title{Get members of unit based on slug}
\usage{
unit_staff(unit_slug = NULL)
}
\arguments{
\item{unit_slug}{string representing KTH unit}
}
\value{
tibble of unit numbers, including kthid
}
\description{
Get members of unit based on slug
}
