% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{wos_attribution}
\alias{wos_attribution}
\title{Generate Web of Science attribution}
\usage{
wos_attribution(year)
}
\arguments{
\item{year}{the year to use for copyright note, current year if not given}
}
\value{
a string with WoS attribution for given copyright year
}
\description{
Generate Web of Science attribution
}
