% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_bullet}
\alias{abm_bullet}
\title{Create bullet graph with reference line}
\usage{
abm_bullet(label, value, reference, roundto = 1, pct = FALSE)
}
\arguments{
\item{value}{the value of the indicator, displayed as a horizontal wide line}

\item{measure}{a label for the indicator, shown to the left of the gauge}

\item{tdarget}{a reference value displayed as a vertical thin line}
}
\value{
a ggplot object
}
\description{
Create bullet graph with reference line
}
