% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_dash_indices}
\alias{abm_dash_indices}
\title{Retrieve dashboard indicators for ABM}
\usage{
abm_dash_indices(data)
}
\arguments{
\item{data}{dataset with publications as tibble}
}
\value{
list with indicator values for dashboard startpage
}
\description{
Retrieve dashboard indicators for ABM
}
