% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{abm_divisions}
\alias{abm_divisions}
\title{Divisions used in ABM}
\usage{
abm_divisions(con = con_bib())
}
\arguments{
\item{con}{database connection, Default: con_bib()}
}
\value{
tibble with divisions
}
\description{
Divisions used in ABM
}
\examples{
\dontrun{
if(interactive()){
 abm_divisions()
 }
}
}
