% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_slug.R
\name{abm_graph_divisions}
\alias{abm_graph_divisions}
\title{ABM divisions as force directed network}
\usage{
abm_graph_divisions(
  base_url = "dash/",
  use_size = FALSE,
  link_encoder = NULL,
  links_excluded = NULL,
  prune_graph = FALSE
)
}
\arguments{
\item{base_url}{pattern to prefigate links from nodes with in JS click action}

\item{use_size}{boolean to indicate whether to size bubbles by publication
volume, default is FALSE}

\item{link_encoder}{function to use for encoding outgoing link identifiers,
default is NULL but can be set to for example 
\code{function(x) URLencode(x, reserved = TRUE)} to transform outgoing links}

\item{links_excluded}{a set of node ids for which links will be excluded, default NULL}

\item{prune_graph}{boolean to indicate if certain non-research nodes should
be removed, default: FALSE}
}
\value{
a force directed network object from NetworkD3
}
\description{
After crawling organizational data, this is used for displaying an 
interactive graph with a force directed network of the units.
}
\examples{
\dontrun{
if(interactive()){
 abm_graph_divisions()
 }
}
}
\seealso{
\code{\link[networkD3]{forceNetwork}},\code{\link[networkD3]{JS}}
}
