% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_oa_data}
\alias{abm_oa_data}
\title{Retrieve OA-data for ABM}
\usage{
abm_oa_data(con = con_bib(), unit_code)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{for filtering on one or more unit code(s), which can be KTH, a one letter school code, an integer department code or a KTH-id (optional)}
}
\value{
tibble with OA-status of all publications from selected organizational units
}
\description{
Retrieve OA-data for ABM
}
