% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_public_data}
\alias{abm_public_data}
\title{Public data from the Annual Bibliometric Monitoring project}
\usage{
abm_public_data(overwrite_cache = FALSE)
}
\arguments{
\item{overwrite_cache}{logical (by default FALSE) specifying whether 
the cache should be refreshed}
}
\value{
a list with two slots - "meta" for organizational unit metadata info 
  and "units" with a named list of results (set of 5 different tibbles for each of the units).
}
\description{
This returns an object which contains data for the various higher 
organizational units at KTH
}
\details{
Data is cached in a local application directory by default and
is returned from there unless the parameter overwrite_cache is TRUE. 
To get fresh data cached, specify this flag.
}
\examples{
\dontrun{

# get all public data from the ABM
public <- abm_public_data()

# get public data specifically for KTH and table 1
unit_kth <- public \%>\% pluck("units", "KTH", 1)

# get summary data for KTH
public \%>\% pluck("units", "KTH", "summaries")
 
# get public data specifically for KTH and table 1
unit_kth <- public \%>\% pluck("units", "KTH", 1)

# get public data for the school "I" and all five tables
unit_i <- public \%>\% pluck("units", "I")

# get public data for the architecture institution, table 1
uc <- public \%>\% pluck("meta") \%>\% 
  filter(unit_long_en == "Architecture") \%>\% pull(unit_code)
  
public \%>\% pluck("units", uc, 1)

}   
}
