% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_publications}
\alias{abm_publications}
\title{Retrieve publication list for ABM}
\usage{
abm_publications(con = con_bib(), unit_code, pub_year)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{unit_code}{the code for the analyzed unit (KTH, a one letter school code, an integer department code or a KTH-id)}

\item{pub_year}{publication year(s) to analyze (optional, assuming master table holds only relevant years)}
}
\value{
tibble with publication list for selected unit
}
\description{
Retrieve publication list for ABM
}
