% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table3}
\alias{abm_table3}
\title{Retrieve Table 3 (Field normalized citations) for ABM}
\usage{
abm_table3(
  data,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year,
  intervals = FALSE
)
}
\arguments{
\item{data}{dataset with publications as tibble}

\item{analysis_start}{first publication year of analysis, default from abm_config()}

\item{analysis_stop}{last publication year of analysis, default from abm_config()}

\item{intervals}{set to TRUE to use 3 year intervals, default FALSE}
}
\value{
tibble with field normalized citations and number/share of top10 publications by 3 year interval
}
\description{
Retrieve Table 3 (Field normalized citations) for ABM
}
