% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table4}
\alias{abm_table4}
\title{Retrieve Table 4 (Journal impact) for ABM}
\usage{
abm_table4(
  data,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year
)
}
\arguments{
\item{data}{dataset with publications as tibble}

\item{analysis_start}{first publication year of analysis, default from abm_config()}

\item{analysis_stop}{last publication year of analysis, default from abm_config()}
}
\value{
tibble with field normalized journal citation score and number/share of publications in top20 journals
}
\description{
Retrieve Table 4 (Journal impact) for ABM
}
