% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table5}
\alias{abm_table5}
\title{Retrieve Table 5 (Co-publishing) for ABM}
\usage{
abm_table5(
  data,
  analysis_start = abm_config()$start_year,
  analysis_stop = abm_config()$stop_year,
  intervals = FALSE
)
}
\arguments{
\item{data}{dataset with publications as tibble}

\item{analysis_start}{first publication year of analysis, default from abm_config()}

\item{analysis_stop}{last publication year of analysis, default from abm_config()}

\item{intervals}{set to TRUE to use 3 year intervals, default FALSE}
}
\value{
tibble with number/share of international copublications and copublications with Swedish non-university organizations
}
\description{
Retrieve Table 5 (Co-publishing) for ABM
}
