% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{abm_table6}
\alias{abm_table6}
\title{Retrieve Table 6 (OA data) for ABM}
\usage{
abm_table6(data)
}
\arguments{
\item{data}{dataset with publications as tibble}
}
\value{
tibble with OA-status of all publications from incoming data
}
\description{
Retrieve Table 6 (OA data) for ABM
}
