% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_staffbased.R
\name{abm_table6_alt}
\alias{abm_table6_alt}
\title{Retrieve Table 6 (OA data) for ABM}
\usage{
abm_table6_alt(con = con_bib(), data)
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}

\item{data}{dataset with publications as tibble, incl PID for joining}
}
\value{
tibble with OA-status of all publications from selected organizational units
}
\description{
Retrieve Table 6 (OA data) for ABM
}
