% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-elements.R
\name{abm_ui_valuebox_publications}
\alias{abm_ui_valuebox_publications}
\title{Value box for number of publications in ABM}
\usage{
abm_ui_valuebox_publications(df, lastyear, vbcolor)
}
\arguments{
\item{df}{data frame with DiVA publications data in a specific format}

\item{lastyear}{last year with DiVA publications}

\item{vbcolor}{color of valueBox}
}
\description{
Value box for number of publications in ABM
}
