% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{ad_search}
\alias{ad_search}
\title{Search Active Directory at KTH using ldapsearch}
\usage{
ad_search(search_term, search_type = c("kthid", "accountname"))
}
\arguments{
\item{search_term}{the KTH id or account name to use in the search query}

\item{search_type}{one of "kthid" or "accountname" for the type of search to make}
}
\value{
tibble with key value pairs
}
\description{
This function uses the ldapsearch utility to query the KTH Active Directory
}
\details{
Besides the openldap utils used for LDAP queries it also requires environment variables to be set in .Renviron (LDAP_USER, LDAP_PASS, LDAP_HOST and LDAP_BASE for the service account used for the queries)

this function requires that the system library openldap is available ("sudo apt install openldap-utils")
}
\examples{
ad_search("markussk", "accountname")
}
