% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_sync}
\alias{db_sync}
\title{Sync the MS SQL Server database BIBMON to a local SQLite3 db}
\usage{
db_sync(
  tables_included,
  tables_excluded = c("Document", "Bestresaddr_KTH", "DIVA_School_Dept",
    "Diva_departments", "Doc_statistics"),
  overwrite_existing = FALSE
)
}
\arguments{
\item{tables_included}{a vector of table names in the source db to be 
included, by default all tables are included except those excluded}

\item{tables_excluded}{a vector of table names in the source db to be
excluded, by default a number of tables are excluded, specify NULL
to not explicitly exclude any tables}

\item{overwrite_existing}{a logical to indicate whether destination tables
should be overwritten if they already exist}
}
\value{
invisible result with vector of boolean status flags for synced tables
}
\description{
This function syncs db tables from an mssql source db and
writes the data into a local SQLite3 db using buffering, with
chunk size set to 1e4 items per chunk, in order to avoid out of
memory exceptions when moving large tables.
}
