% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{get_pt_ordning}
\alias{get_pt_ordning}
\title{Get order of publication type for ABM table 1}
\usage{
get_pt_ordning(con = con_bib())
}
\arguments{
\item{con}{connection to db, default is to use mssql connection}
}
\value{
tibble with pt_ordning and diva_publiation_type
}
\description{
Get order of publication type for ABM table 1
}
