% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{run_api}
\alias{run_api}
\title{Running this function launches a Plumber API presenting
resources bundled in this package}
\usage{
run_api(teardown = FALSE)
}
\arguments{
\item{teardown}{logical indicating if the API should be shut down if running}
}
\description{
Running this function launches a Plumber API presenting
resources bundled in this package
}
