% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{run_app}
\alias{run_app}
\title{Running this function launches a Shiny web application presenting
the datasets bundled in this package}
\usage{
run_app(example)
}
\arguments{
\item{example}{the name of the bundled Shiny application example}
}
\description{
Running this function launches a Shiny web application presenting
the datasets bundled in this package
}
