% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{unit_info}
\alias{unit_info}
\title{Retrieve information about ABM units (level 0-2) from database or from package data}
\usage{
unit_info(con)
}
\arguments{
\item{con}{connection to db}
}
\value{
tibble with information about ABM units
}
\description{
If a database connection is given, abm_org_info is read from database,
otherwise abm_public_kth$meta is returned
}
