% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{wos_disclaimer}
\alias{wos_disclaimer}
\title{Generate Web of Science disclaimer}
\usage{
wos_disclaimer(year)
}
\arguments{
\item{year}{the year to use for copyright note, current year if not given}
}
\value{
a string with WoS disclaimer for given year
}
\description{
Generate Web of Science disclaimer
}
